[FBZ10350] Refactor and use libraries update

+++ cookbooks/ey-app-logs/recipes/default.rb
@@ -32,6 +32,7 @@ existing_apps.each do |existing_app|
   unless node["dna"]["applications"].include? existing_app
     execute "Remove files of detached apps" do
       command %(rm -rf /var/log/engineyard/apps/#{existing_app})
+      not_if { ::Dir.glob("/data/#{existing_app}/release*").nil? }
     end
   end
 end
